requirements-dev.txt
coverage==3.7.1
pytest==2.6.4
pytest-cov==1.8.0
Sphinx==1.2.3
mock==1.0.1
Flask-WebTest==0.0.7